refactor: 💡 Migrate MaskedTextField to MUI5
✅ Closes: #732
